TDM Spawn, Respawn, Health Packs, and Health Pools
Instantiates Boat
has 2 prefabs and game objects, one to damage (just for testing) and
one for healing (actually will be implemented)
Respawns boat after 8 seconds, can be changed
Logs lives of team and logs when one of the teams wins.
Game object Team 1 Spawner controls the spawning of a certain team with
its attached script. Also is parent to the locations of the spawn
points for the team

Created scripts Health (to manage the health and death state of a
player), HealthPack (not much, just for the despawning and respawning
of the health pack), Master (Controls amount of teams and the lives
each team has), and SpawnScript (controls spawning and respawning of
player(s) at given locations