Added Simple Buoyancy and Wave Script
Right now the buoyancy and waves don't interact, as the buoyancy only
checks for the object's position relative to y = 0. I don't know how to
make it interact with the mesh of the water, or if the waves I added are
even the kind we want. I'll talk to design on Friday and see what
they're thinking for waves.