StatsBoat and Canvas Prefabs (and new Scene)
The _CrewManagement Scene and StatsBoat prefab are for testing new
stats-related scripts without messing with the current boat and scene.
The Canvas prefab at the moment contains the Crew Management User
Interface. This UI is composed of 3 buttons that the user can click to
modify their stats (currently using placeholder images).