Networked Ships + Shooting
Created Networked versions of many scripts and objects to allow for
rough multiplayer.
To use, build and run one copy and run another copy from within the
editor. Choose "Host" on one and "Client" on the other. For the same
machine, leave option as "localhost", otherwise change it to your local/
external IP address.