Added toon shaders + trireme + human ship 1/2
Due to the lack of models and design documentation on how ship selection
should work, I've just added the foundations for future models to be
imported as well as imported the functional models that work (albeit
properties are offscaled and blown up). Need further discussion with art
team to progress further with importing new models.