First attempt at connecting Crew Management to game
The Crew Management System has been reworked into a multiplier system
with four positive and four negative stages. The values for the stages
are set to (in order from stage -4 to stage +4): 0.25, 0.33, 0.5, 0.66,
0, 1.5, 2.0, 2.5, and 3.0. The scripts that this system influences are
BoatMovementNetworked, BroadsideCannonFireNetworked, and Health.

The CrewManagementUI has been altered to work with this new system, and
the UI buttons have been removed entirely in favor of hotkeys only
(Currently set to 1, 2, and 3).

Details about the changes can be found in the comments of the scripts
listed above.

BUG: The variables created in the other scripts, the ones the Crew
Management Script changes, do influence calculations when set to a value
in the code. However, when a player chooses to alter a stat during
gameplay, the stat change is not reflected. To clarify, when a player
chooses to change a stat, the respective variable is updated properly in
the inspector and debug log, but the game acts as if the change never
happened when using the variable. Does this mean the Crew Management
System has to be networked, or is there an error with how the variables
are being updated across scripts?

Due to this BUG, I will be commiting these changes but NOT pushing until
this is sorted out.