Added Health Packs
Added health packs that spawn every x seconds. Health packs spawn where
the "HealthPack Spawner" empty gameobject is placed, and restore health
on pick up. The spawn time is set in the HealthPack Spawner object, and
the amount of health restored is set in Health.cs in the ship object.
All changes are present in the scene healthPackTest.