Ship Selection in Lobby (with some networking)
Added:
- Scene called Lobby with Ship Select for playtesting; it has the
ability for a client to select a ship out of a list
- Added a feature where the player can rotate a 3D model of their ship
with the mouse once selecting it (RotateWithMouse.cs)
- The scene is called instead of "Lobby" when hosting and joining a
game.

Modified:
- LobbyManager.cs- Added ship select function
- MultiplayerManager.cs- Modified SpawnMessage and added a
localShipSelect variable
- PlayerIconController.cs- Added commands and rpc calls for ship select
- Player.cs- added ship instance variable

Comments: KYLE A.-- The ship selection does not reflect entirely to
connected players due to a null exception in the rpc call. This could be
because of the bugs you were ironing out over this week. I also need
more clarification of how your PlayerIconController works, but we can
talk about that this Friday.