Collision Work on Bramble Projectile - part 2
>Projectile now correctly recognizes when it has collided with a boat.
This is done by checking the layer of each object the projectile goes
through using an OnTriggerEnter function, where other is a variable
standing in for the collided object. The layer currently being checked
for is 8, this was found to be the numerical value placed on "Boat" in
layers.
>Further enlarged the projectile collider.
>Set the collider on the projectile to be a trigger, adressing many of
the issues seen earlier.

Further work will be investiagting how to dsicern between the owning
player and all other players. Also further work will be done on
instiagting damage to hit enemy players (teams must also be able to be
discerend)