Bramble projectile no longer delays at max distance
The bramble projectile should now feel more natural. It no longer
stutter or delays at the max distance point.

>var Speed is not longer public, and is now set to a constant of 1
>speed is now set to the value (The distance from the firing point to
the max distance / how long we want it to take to get there

With these changes the projectile show need more fluid. However, speed
or rate changes to the projectile must now be done on the public
travelTime variable. The code now adapts based on how long we want it to
take to get to that point.