Sunship Heavy Weapon BASE functionality
#Joseph Cash#
Sunship now shoots a projectile in the forward direction (relative).
The projectile checks the distance between its fired point and its
current location, once the (modifiable) threshhold has been met: it will
destroy its self.

Further work needs to be done to create the AOE effect. We need to talk
to programming on how to make this happen.