Weapon System Enhancements
The colliders on the cannons of the prefab were disabled because the
cannonballs were colliding with the cannons when they were instantiated
and there is no reason for them to have colliders anyways.
BoatMovementNetworked was changed so the boats will work with any number
of cannons and swivel guns. SwivelGun was changed to extend
BroadsideCannonNetworked so the only difference between a broadside
cannon and a swivel gun is which of the two scripts is attached.