Camera settings storage and values range Change
Added funcitonality where camera settings (sensitivity and inversions)
will ve stored between sessions using PlayerPrefs. Also, in a new scene
called MainLevelKyleCameraSettingsChanged (!) changed the value ranged
for the camera settings to avoid being able to set sensitivity to 0 (for
vertical and horizontal) and increasing the floor for scroll sensitivity
(1->5) which is more reasonable.